Revert "[FL-3896] Split BadUSB into BadUSB and BadBLE (#3931)"
This reverts commit 0eaad8bf64f01a6f932647a9cda5475dd9ea1524.
