--- applications/system/bad_ble/application.fam
@@ -1,12 +1,0 @@
-App(
-    appid="bad_ble",
-    name="Bad BLE",
-    apptype=FlipperAppType.EXTERNAL,
-    entry_point="bad_ble_app",
-    stack_size=2 * 1024,
-    icon="A_BadUsb_14",
-    fap_libs=["assets", "ble_profile"],
-    fap_icon="icon.png",
-    fap_icon_assets="assets",
-    fap_category="Bluetooth",
-)
