--- applications/system/bad_ble/bad_ble_app.h
@@ -1,11 +1,0 @@
-#pragma once
-
-#ifdef __cplusplus
-extern "C" {
-#endif
-
-typedef struct BadBleApp BadBleApp;
-
-#ifdef __cplusplus
-}
-#endif
