--- applications/system/bad_ble/scenes/bad_ble_scene.h
@@ -1,29 +1,0 @@
-#pragma once
-
-#include <gui/scene_manager.h>
-
-// Generate scene id and total number
-#define ADD_SCENE(prefix, name, id) BadBleScene##id,
-typedef enum {
-#include "bad_ble_scene_config.h"
-    BadBleSceneNum,
-} BadBleScene;
-#undef ADD_SCENE
-
-extern const SceneManagerHandlers bad_ble_scene_handlers;
-
-// Generate scene on_enter handlers declaration
-#define ADD_SCENE(prefix, name, id) void prefix##_scene_##name##_on_enter(void*);
-#include "bad_ble_scene_config.h"
-#undef ADD_SCENE
-
-// Generate scene on_event handlers declaration
-#define ADD_SCENE(prefix, name, id) \
-    bool prefix##_scene_##name##_on_event(void* context, SceneManagerEvent event);
-#include "bad_ble_scene_config.h"
-#undef ADD_SCENE
-
-// Generate scene on_exit handlers declaration
-#define ADD_SCENE(prefix, name, id) void prefix##_scene_##name##_on_exit(void* context);
-#include "bad_ble_scene_config.h"
-#undef ADD_SCENE
