--- applications/system/bad_ble/scenes/bad_ble_scene_config.h
@@ -1,7 +1,0 @@
-ADD_SCENE(bad_ble, file_select, FileSelect)
-ADD_SCENE(bad_ble, work, Work)
-ADD_SCENE(bad_ble, error, Error)
-ADD_SCENE(bad_ble, config, Config)
-ADD_SCENE(bad_ble, config_layout, ConfigLayout)
-ADD_SCENE(bad_ble, confirm_unpair, ConfirmUnpair)
-ADD_SCENE(bad_ble, unpair_done, UnpairDone)
